refactor(sidecar): split sidecar.go into focused files (#1155)
Extract independent function groups from the monolithic sidecar.go
(3879 LoC) into six new files:

- events.go:    HandleEvent() + 12 handle* methods (event dispatch)
- state.go:     upsertState(), writeStateChange*(), writeEvent(),
                writeStartupError(), cancelIdleTimer(),
                cancelRecoveryTimer(), currentDBState()
- notify.go:    notifyCoordinator(), notifyParentWorkerOnStartupFailure(),
                reviewAgentParentSession(), validateOrRefreshCoordinatorSID(),
                deliverNotificationViaHTTP(), buildNotifyPromptBody()
- dashboard.go: pushDashboardEvent(), dashboardSocketPath(),
                touchDashboardSentinel()
- helpers.go:   strPtr(), truncate(), marshalTruncated(),
                isHighImpactCommand(), extractBashCommand(),
                extractMessageIDFromPayload(), repoFromSession(),
                isCoordinator(), isCoordinatorSession(),
                isHostAPITerminalState(), isReviewAgentSession(),
                parseSpawnSessionName()
- host_api.go:  hostAPIHandler(), hostAPIServeLogsTail(),
                hostAPIServeLogsFollow(), worktreePathForSession(),
                isKnownReviewAgent(), knownReviewAgentNames()

sidecar.go retains: package doc, Sidecar struct, Config, Clock/Timer
interfaces, RealClock, New(), containerMgr, Run(), runStartupHTTP(),
runStartupStdio(), Shutdown(), and the top-level consts/vars they depend on.

Pure structural refactor — no behaviour changes, no signature changes.
